requirements.txt
fastapi==0.95.1
uvicorn==0.20.0
tensorflow==2.10.0
numpy==1.23.4
pillow==9.3.0